логика сортировки вынесена в отдельный хук
