+++ packages/client/src/components/organisms/leaderboard/leaderboard.component.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState, memo } from 'react'
+import { memo } from 'react'
 import {
   Control,
   Item,
@@ -13,11 +13,11 @@ import {
   SortElement,
   SortElementIcon,
 } from './leaderboard.styles'
+import { H2 } from '@/global-styles'
 import { Input } from '@/components/molecules'
 import IconSort from '@/assets/icons/sort.svg'
 import { LeaderboardElementProps } from './leaderboard-types'
-import mergeSort from '@/utils/mergeSort'
-import { H2 } from '@/global-styles'
+import useLeaderboard from './leaderboard.logics'
 
 import avatarDefault from '@/assets/images/avatarDefault.png'
 
@@ -42,43 +42,15 @@ const LeaderboardElement = memo((props: LeaderboardElementProps) => {
 })
 
 export const Leaderboard = ({ data }: { data: LeaderboardElementProps[] }) => {
-  const [search, setSearch] = useState('')
-  const [nameSort, setNameSort] = useState('DESC')
-  const [positionSort, setPositionSort] = useState('DESC')
-  const [sortData, setSortData] = useState(data)
-  const [noResults, setNoResults] = useState(false)
-
-  useEffect(() => {
-    let filteredData = [...data]
-    if (search !== '') {
-      filteredData = data.filter(el => el.name.indexOf(search) !== -1)
-      setNoResults(filteredData.length === 0)
-    }
-    const sortedData = mergeSort(filteredData, 'position', 'DESC')
-    setSortData(sortedData)
-    setPositionSort('ASC')
-  }, [data, search])
-
-  useEffect(() => {
-    setSortData(prevSortData => {
-      const sortedData = mergeSort(prevSortData, 'name', nameSort)
-      return sortedData
-    })
-  }, [nameSort])
-
-  useEffect(() => {
-    setSortData(prevSortData => {
-      const sortedData = mergeSort(prevSortData, 'position', positionSort)
-      return sortedData
-    })
-  }, [positionSort])
-
-  const dataWithPositions = data.map((el, index) => ({ ...el, position: index + 1 }))
-
-  const sortedData = sortData.map((el) => {
-    const position = dataWithPositions.find((e) => e.name === el.name)?.position
-    return { ...el, position }
-  })
+  const {
+    setSearch,
+    nameSort,
+    setNameSort,
+    positionSort,
+    setPositionSort,
+    sortedData,
+    noResults,
+  } = useLeaderboard({ data })
 
   return (
     <>
